+++ javacutil/src/main/java/org/checkerframework/javacutil/TreeUtils.java
@@ -2744,8 +2744,21 @@ public final class TreeUtils {
    *
    * @param tree a tree
    * @return true if {@code tree} is a method reference with a raw type to the left of {@code ::}
+   * @deprecated use {@link #isRawTypedMemberReference(ExpressionTree)}
    */
+  @Deprecated // 2026-08-31
   public static boolean isLikeDiamondMemberReference(ExpressionTree tree) {
+    return isRawTypedMemberReference(tree);
+  }
+
+  /**
+   * Returns true if {@code tree} is a method reference with a raw type to the left of {@code ::}.
+   * For example, {@code Class::getName}.
+   *
+   * @param tree a tree
+   * @return true if {@code tree} is a method reference with a raw type to the left of {@code ::}
+   */
+  public static boolean isRawTypedMemberReference(ExpressionTree tree) {
     if (!(tree instanceof MemberReferenceTree memberReferenceTree)) {
       return false;
     }
@@ -2764,7 +2777,7 @@ public final class TreeUtils {
    */
   public static boolean needsTypeArgInference(MemberReferenceTree memberReferenceTree) {
     if (isDiamondMemberReference(memberReferenceTree)
-        || isLikeDiamondMemberReference(memberReferenceTree)) {
+        || isRawTypedMemberReference(memberReferenceTree)) {
       return true;
     }